Move project model tests to existing file
